feat: port Joy of Energy JS/React to CLJS/Reagent
WF-327

Visually the port should be 1:1, but the implementation details have
been slightly modified.  Some of the implementation modifications are
to make the code idiomatic CLJS and other changes are just to make the
code maintaininable and easier to understand.  Not that I didn't go as
far as I would like with the implementation changes.  We can layer on
further changes as we go.